read receipts and personal token as service
