+++ imports/message-read-receipt/server/api/methods/getReadReceipts.js
@@ -1,10 +1,9 @@
 import { Meteor } from 'meteor/meteor';
 
-import { ReadReceipt } from '../../lib/ReadReceipt';
-
 Meteor.methods({
-	getReadReceipts({ messageId }) {
-		if (!Meteor.userId()) {
+	async getReadReceipts({ messageId }) {
+		const uid = Meteor.userId();
+		if (!uid) {
 			throw new Meteor.Error('error-invalid-user', 'Invalid user', { method: 'getReadReceipts' });
 		}
 
@@ -12,17 +11,6 @@ Meteor.methods({
 			throw new Meteor.Error('error-invalid-message', 'The required \'messageId\' param is missing.', { method: 'getReadReceipts' });
 		}
 
-		const message = RocketChat.models.Messages.findOneById(messageId);
-
-		if (!message) {
-			throw new Meteor.Error('error-invalid-message', 'Invalid message', { method: 'getReadReceipts' });
-		}
-
-		const room = Meteor.call('canAccessRoom', message.rid, Meteor.userId());
-		if (!room) {
-			throw new Meteor.Error('error-invalid-room', 'Invalid room', { method: 'getReadReceipts' });
-		}
-
-		return ReadReceipt.getReceipts(message);
+		return RocketChat.Services.call('messageReadReceipt.getReadReceipts', { uid, messageId });
 	},
 });
