+++ imports/personal-access-tokens/server/api/actions/index.js
@@ -1,0 +1,9 @@
+import generateToken from './generateToken';
+import regenerateToken from './regenerateToken';
+import removeToken from './removeToken';
+
+export default {
+	...generateToken,
+	...regenerateToken,
+	...removeToken,
+};
