+++ imports/services/index.js
@@ -3,6 +3,9 @@ import Queue from './services/queue.js';
 import Authorization from './services/authorization';
 import User from './services/user';
 import Settings from './services/settings';
+import Chat from './services/chat';
+import PersonalAccessTokens from '../personal-access-tokens/server/service';
+import GetReadReceipts from '../message-read-receipt/server/service';
 import { ServiceBroker } from 'moleculer';
 import { Meteor } from 'meteor/meteor';
 import { RocketChat } from 'meteor/rocketchat:lib';
@@ -18,6 +21,9 @@ broker.createService(Queue);
 broker.createService(Authorization);
 broker.createService(User);
 broker.createService(Settings);
+broker.createService(Chat);
+broker.createService(PersonalAccessTokens);
+broker.createService(GetReadReceipts);
 
 RocketChat.Services = broker;
