+++ packages/rocketchat-lib/server/methods/sendMessage.js
@@ -1,14 +1,12 @@
 import { Meteor } from 'meteor/meteor';
 import { check } from 'meteor/check';
-import { Random } from 'meteor/random';
-import { TAPi18n } from 'meteor/tap:i18n';
-import moment from 'moment';
 
 Meteor.methods({
-	sendMessage(message) {
+	async sendMessage(message) {
 		check(message, Object);
+		const uid = Meteor.userId();
 
-		if (!Meteor.userId()) {
+		if (!uid) {
 			throw new Meteor.Error('error-invalid-user', 'Invalid user', {
 				method: 'sendMessage',
 			});
@@ -18,74 +16,7 @@ Meteor.methods({
 			throw new Error('The \'rid\' property on the message object is missing.');
 		}
 
-		if (message.ts) {
-			const tsDiff = Math.abs(moment(message.ts).diff());
-			if (tsDiff > 60000) {
-				throw new Meteor.Error('error-message-ts-out-of-sync', 'Message timestamp is out of sync', {
-					method: 'sendMessage',
-					message_ts: message.ts,
-					server_ts: new Date().getTime(),
-				});
-			} else if (tsDiff > 10000) {
-				message.ts = new Date();
-			}
-		} else {
-			message.ts = new Date();
-		}
-
-		if (message.msg) {
-			const adjustedMessage = RocketChat.messageProperties.messageWithoutEmojiShortnames(message.msg);
-
-			if (RocketChat.messageProperties.length(adjustedMessage) > RocketChat.settings.get('Message_MaxAllowedSize')) {
-				throw new Meteor.Error('error-message-size-exceeded', 'Message size exceeds Message_MaxAllowedSize', {
-					method: 'sendMessage',
-				});
-			}
-		}
-
-		const user = RocketChat.models.Users.findOneById(Meteor.userId(), {
-			fields: {
-				username: 1,
-				name: 1,
-			},
-		});
-
-		const room = Meteor.call('canAccessRoom', message.rid, user._id);
-		if (!room) {
-			return false;
-		}
-
-		const subscription = RocketChat.models.Subscriptions.findOneByRoomIdAndUserId(message.rid, Meteor.userId());
-		if (subscription && (subscription.blocked || subscription.blocker)) {
-			RocketChat.Notifications.notifyUser(Meteor.userId(), 'message', {
-				_id: Random.id(),
-				rid: room._id,
-				ts: new Date,
-				msg: TAPi18n.__('room_is_blocked', {}, user.language),
-			});
-			throw new Meteor.Error('You can\'t send messages because you are blocked');
-		}
-
-		if ((room.muted || []).includes(user.username)) {
-			RocketChat.Notifications.notifyUser(Meteor.userId(), 'message', {
-				_id: Random.id(),
-				rid: room._id,
-				ts: new Date,
-				msg: TAPi18n.__('You_have_been_muted', {}, user.language),
-			});
-			throw new Meteor.Error('You can\'t send messages because you have been muted');
-		}
-
-		if (message.alias == null && RocketChat.settings.get('Message_SetNameToAliasEnabled')) {
-			message.alias = user.name;
-		}
-
-		if (Meteor.settings.public.sandstorm) {
-			message.sandstormSessionId = this.connection.sandstormSessionId();
-		}
-
-		RocketChat.metrics.messagesSent.inc(); // TODO This line needs to be moved to it's proper place. See the comments on: https://github.com/RocketChat/Rocket.Chat/pull/5736
-		return RocketChat.sendMessage(user, message, room);
+		return RocketChat.Services.call('chat.sendMessage', { uid, message });
 	},
 });
 // Limit a user, who does not have the "bot" role, to sending 5 msgs/second
